move VersionForm and VersionFields to /components/form/version
